make more robust; clean up

+++ ecdh25519/test.c
@@ -84,7 +84,7 @@ static void run_speed(void)
   (void)cycles;
   sprintf(cycles_str, "[cycle counts not meaningful in qemu emulation]\n");
 #else
-  sprintf(cycles_str, "%llu\n", (unsigned long long)cycles);
+  sprintf(cycles_str, "%llu\n", cycles);
 #endif
   hal_send_str(cycles_str);
 
@@ -96,7 +96,7 @@ static void run_speed(void)
   (void)cycles;
   sprintf(cycles_str, "[cycle counts not meaningful in qemu emulation]\n");
 #else
-  sprintf(cycles_str, "%llu\n", (unsigned long long)cycles);
+  sprintf(cycles_str, "%llu\n", cycles);
 #endif
   hal_send_str(cycles_str);
