+++ ml-dsa/test.c
@@ -226,7 +226,7 @@ static void run_speed(void)
     (void)cycles;
     sprintf(cycles_str, "[cycle counts not meaningful in qemu emulation]\n");
 #else
-    sprintf(cycles_str, "%llu\n", (unsigned long long)cycles);
+    sprintf(cycles_str, "%llu\n", cycles);
 #endif
     hal_send_str(cycles_str);
 
@@ -239,7 +239,7 @@ static void run_speed(void)
     (void)cycles;
     sprintf(cycles_str, "[cycle counts not meaningful in qemu emulation]\n");
 #else
-    sprintf(cycles_str, "%llu\n", (unsigned long long)cycles);
+    sprintf(cycles_str, "%llu\n", cycles);
 #endif
     hal_send_str(cycles_str);
 
@@ -252,7 +252,7 @@ static void run_speed(void)
     (void)cycles;
     sprintf(cycles_str, "[cycle counts not meaningful in qemu emulation]\n");
 #else
-    sprintf(cycles_str, "%llu\n", (unsigned long long)cycles);
+    sprintf(cycles_str, "%llu\n", cycles);
 #endif
     hal_send_str(cycles_str);
 
@@ -265,7 +265,7 @@ static void run_speed(void)
     (void)cycles;
     sprintf(cycles_str, "[cycle counts not meaningful in qemu emulation]\n");
 #else
-    sprintf(cycles_str, "%llu\n", (unsigned long long)cycles);
+    sprintf(cycles_str, "%llu\n", cycles);
 #endif
     hal_send_str(cycles_str);
